+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [1.5.0] - Unreleased
 
+* Improved the data labels in pie charts
+* Improved data comparison UX to list the columns that are extra
 * Bugs
     * Data comparison of two sections would say the data differs even when there are 0 different rows or columns.
     * Removed 1 pixel of unwanted gap between the top of the tabular results and the top border. It made it possible to see the text that was being scrolled, super annoying.